requirements/base.txt
accelerate==1.10.1
alabaster==1.0.0
anyio==4.10.0
argon2-cffi==23.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.9.1
asttokens==3.0.0
babel==2.17.0
boto3==1.40.21
botocore==1.40.21
certifi==2025.8.3
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.3
click==8.2.1
Collectfasta==3.2.1
colorama==0.4.6
coverage==7.8.0
crispy-bootstrap5==2025.4
cryptography==44.0.3
cssbeautifier==1.15.4
decorator==5.2.1
distlib==0.4.0
Django==5.2.5
django-allauth==65.7.0
django-anymail==13.0
django-appconf==1.1.0
django-compressor==4.5.1
django-coverage-plugin==3.1.0
django-crispy-forms==2.3
django-debug-toolbar==5.1.0
django-environ==0.12.0
django-extensions==4.0
django-model-utils==5.0.0
django-redis==5.4.0
django-storages==1.14.6
django-stubs==5.1.3
django-stubs-ext==5.2.2
djlint==1.36.4
docutils==0.21.2
EditorConfig==0.17.1
einops==0.8.1
executing==2.2.1
factory_boy==3.3.2
faiss-cpu==1.12.0
Faker==37.6.0
fido2==1.2.0
filelock==3.19.1
fsspec==2025.9.0
gunicorn==23.0.0
h11==0.16.0
hf-xet==1.1.9
hiredis==3.1.0
huggingface-hub==0.34.4
identify==2.6.13
idna==3.10
imagesize==1.4.1
iniconfig==2.1.0
ipdb==0.13.13
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
jsbeautifier==1.15.4
json5==0.12.1
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mpmath==1.3.0
mypy==1.15.0
mypy_extensions==1.1.0
networkx==3.5
nodeenv==1.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.1.105
packaging==25.0
parso==0.8.5
pathspec==0.12.1
pdfminer.six==20250506
pdfplumber==0.11.7
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.4.0
pluggy==1.6.0
pre_commit==4.2.0
prompt_toolkit==3.0.52
psutil==7.0.0
psycopg==3.2.6
psycopg-binary==3.2.6
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.2
PyPDF2==3.0.1
pypdfium2==4.30.0
pytest==8.3.5
pytest-django==4.11.1
pytest-sugar==1.0.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
PyYAML==6.0.2
qrcode==8.2
rcssmin==1.1.2
redis==5.2.1
regex==2025.9.1
requests==2.32.5
rjsmin==1.2.2
roman-numerals-py==3.1.0
ruff==0.11.4
s3transfer==0.13.1
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
sentence-transformers==2.6.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
snowballstemmer==3.0.1
Sphinx==8.2.3
sphinx-autobuild==2024.10.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.47.3
sympy==1.14.0
termcolor==3.1.0
text-unidecode==1.3
threadpoolctl==3.6.0
tokenizers==0.22.0
torch==2.2.2+cu121
torchaudio==2.2.2+cu121
torchvision==0.17.2+cu121
tqdm==4.67.1
traitlets==5.14.3
transformers==4.56.0
triton==3.4.0
types-PyYAML==6.0.12.20250822
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
virtualenv==20.34.0
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3